Fixed Pistol Burst Fire and Restored Legacy Fire for Minigun
Also added Customize ammo for Freezerthrower

Now in more detail: I noticed that the animation of the pistol during burst fire was broken not only for the usual one, but also for the one with a silencer, I decided to correct this misunderstanding. Also, I did not add the code with the original type of minigun fire.